+++ Tests/Model/JobTimingTest.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Dtc\QueueBundle\Tests\Model;
+
+use Dtc\QueueBundle\Tests\GetterSetterTrait;
+use PHPUnit\Framework\TestCase;
+
+class JobTimingTest extends TestCase
+{
+    use GetterSetterTrait;
+
+    public function testGettersSetters()
+    {
+        $this->runGetterSetterTests('\Dtc\QueueBundle\Model\JobTiming');
+    }
+}